requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
gunicorn==21.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.9
pytz==2024.1
setuptools==69.4.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0